# Speed up OP_REVERSEBYTES test significantly