Summary: Currently, the test for OP_REVERSEBYTES takes ~40s to finish.
The bottleneck is that both the stack item sizes and the flag set are
checked exhaustively combinatorically. This diff still checks both
exhaustively, but separately, and uses a curated set of flags/test cases
for each, respectively. This reduces the test execution time to around ~1s.

W/O the patch:

```
$> time src/test/test_bitcoin -t op_reversebytes_tests
Running 1 test case...

*** No errors detected

real    0m31,690s
user    0m31,663s
sys     0m0,028s
```

W/ the patch:

```
$> time src/test/test_bitcoin -t op_reversebytes_tests
Running 3 test cases...

*** No errors detected

real    0m1,185s
user    0m1,153s
sys     0m0,032s
```